chore: add v2 api folder structure
