+++ bc_obps/registration/api/__init__.py
@@ -1,32 +1,4 @@
-# ruff: noqa: F401
-from . import (
-    business_structures,
-    e2e_test_setup,
-    naics_codes,
-    operators,
-    regulated_products,
-    user_operator,
-    reporting_activities,
-    users,
-    operations,
-)
-from ._operations import operation_id
-from ._operations._operation_id import update_status
-
-from .operator import operator_access_declined, operator_has_admin
-from .user import user_profile, user_app_role
-
-# ruff: noqa: F811
-from .user_operator import (
-    operator,
-    operator_from_user,
-    is_approved_admin_user_operator,
-    user_operator_from_user,
-    user_operator_id,
-    user_operator_id_from_user,
-    user_operator_initial_requests,
-    user_operator_list_from_user,
-    update_status,
-)
-from .select_operator import request_access, request_admin_access
-from .api_base import router  # django_ninja router object
+# ruff: noqa: F403
+from .v1 import *
+from .v2 import *
+from .router import router  # django_ninja router object  # noqa: F401
