--- bc_obps/registration/api/select_operator/request_access.py
@@ -1,26 +1,0 @@
-from registration.api.utils.current_user_utils import get_current_user_guid
-from registration.decorators import authorize, handle_http_errors
-from registration.schema import (
-    SelectOperatorIn,
-    Message,
-    RequestAccessOut,
-)
-from registration.api.api_base import router
-
-from registration.models import (
-    UserOperator,
-)
-from service.error_service.custom_codes_4xx import custom_codes_4xx
-from service.application_access_service import ApplicationAccessService
-
-
-@router.post(
-    "/select-operator/request-access",
-    response={201: RequestAccessOut, custom_codes_4xx: Message},
-    url_name="request_access",
-)
-@authorize(["industry_user"], UserOperator.get_all_industry_user_operator_roles(), False)
-@handle_http_errors()
-def request_access(request, payload: SelectOperatorIn):
-
-    return 201, ApplicationAccessService.request_access(payload.operator_id, get_current_user_guid(request))
