+++ bc_obps/registration/api/v2/__init__.py
@@ -1,0 +1,2 @@
+# ruff: noqa: F401
+from . import sample
